requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pymysql>=1.1.1
pydantic-settings>=2.6.1
PyJWT>=2.10.1
